Add "resultAvailable" function.
resultAvailable works like getResult but it does not throw exception.
getResult throws an exception if the result file does not exist. It is not an error
if getResult is called soon after the launching of the job and there was not enough time
to create the file.
resultAvailable does not throw exceptions. It returns "False" if the result file does
not exist.